+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Dialog for adding a new channel.
+
 ## [0.13.0] - 2024-02-25
 
 ### Added